fix: explicitly specify azure ccm imageTag (#1421)
After upgrading k0s to v1.32.3, the Azure Cloud Controller Manager fails
to deploy due to a limitation in the Helm chart which determines the CCM image version.
It does not currently handle Kubernetes version 1.32:
https://github.com/Mirantis/helm-charts/blob/main/mirantis/cloud-provider-azure/templates/cloud-provider-azure.yaml#L22

As a result, .cloudProviderAzureVersion remains unset, which causes the generated image path to be empty:
https://github.com/Mirantis/helm-charts/blob/main/mirantis/cloud-provider-azure/templates/_image.tpl#L7

When the imageTag is specified, the image path is generated properly.

Fixes #1420