fix: unify infrastructure providers naming (#1440)
There was an inconsistency in infrastructure providers' naming
(for example, searched for `infrastructure-aws` in one place and `aws` in another),
which breaks the removal of the blocking finalizer from the ClusterDeployment.

Now, the full provider name with the infrastructure prefix is being used.

Also, added indexer for the PluggableProvider object to fetch all the exposed infrastructure providers.

Fixes #1438