feat: automatically update stuck sveltos tokens (#1588)
Adds new controller that determines if there is a stuck
unhealthy management SveltosCluster due to expired token
and updates the corresponding secret data with a new token,
so the other sveltos controllers can proceed theirs logic
working with the new kubeconfig.

Bumps binary assets data to 1.32 to be on par with
the current envtest k8s version in use.