ci: improve e2e tests run and cache mgmt
Improves management of caches, using caches where it actually is
required. Usage of cache on self-hosted runner is disabled as well as
cleaning go-related caches after running e2e tests.

Checks disk space if remote provider is presented within the
configuration and decides whether to use cache or not, and removes all
of the available for branch caches.

Runs e2e test jobs only if they are actually required to run: skips
on-prem if no vsphere is presented, and cloud if no cloud providers have
been given.

Fixes release WF to skip *-rc tags in goreleaser.