feat: child telemetry tracker (#1783)
Reworks the current tracker so it now uses a generic interface to call
different types of collectors, calling collection once with a given
interval with some jitter to decrease potential IO pressure.

Adds new child telemetry collection tracker for segment.io.

CLD Event creation still uses the package-level client but does nothing
if it is null, which is the case if telemetry is either disabled or in
the local mode.