feat(restore): support regions restoration (#2073)
* feat(restore): support regions restoration

Adds logic to the cld controller to skip reconciliation until the
related Region has the Pause annotation.

Adds regions watcher to trigger on the Pause annotation deletion.

Reverts changes in the backup controller to fail if a regional backup
fails to be created.

Creates backups for all of the regions, setting errors to the according
regional statuses.

Fixes the restoration path to correctly restore the latest backup
statuses for (not yet) existing region objects.